backend/requirements.txt
# Backend dependencies for FastAPI + PostgreSQL + Gemini

# FastAPI y ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Security & JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic[email]==2.5.0

# Google Gemini
google-generativeai==0.3.0

# Environment variables
python-dotenv==1.0.0

# Utilities
requests==2.31.0
aiofiles==23.2.1

# Logging and monitoring
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development (opcional)
# black==23.12.0
# flake8==6.1.0
# mypy==1.7.1
